refactor: menu items hooks folder
